pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Nikolay Leonovets <work.nleonovec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
fastapi = {extras = ["all"], version = "^0.104.1"}
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
greenlet = "^3.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
